feat: remove dep on ethers
