+++ package.json
@@ -37,7 +37,6 @@
     "@scure/bip39": "^1.2.1",
     "@walletconnect/legacy-client": "^2.0.0",
     "@walletconnect/sign-client": "^2.8.0",
-    "ethers": "^6.9.1",
     "lodash-es": "^4.17.21"
   },
   "devDependencies": {
@@ -65,7 +64,6 @@
     "degit": "^2.8.4",
     "eslint": "^8.40.0",
     "eslint-config-prettier": "^8.8.0",
-    "ethers": "^6.9.1",
     "glob": "^10.2.3",
     "json-schema-to-typescript": "^13.1.1",
     "lodash-es": "^4.17.21",
