+++ src/codec/index.ts
@@ -4,5 +4,10 @@ export * from "@scure/base";
 export { resolveBech32Address, translateEthToBech32Address } from "./address";
 export { ethhex } from "./ethhex";
 export { resolveKeyPair } from "./key";
-export { hashEthArbitraryMessage, serialiseSignDoc } from "./serialise";
-export { recoverPubKeyFromEthSignature, signAmino, signDirect } from "./sign";
+export { serialiseSignDoc } from "./serialise";
+export {
+  hashEthArbitraryMessage,
+  recoverPubKeyFromEthSignature,
+  signAmino,
+  signDirect,
+} from "./sign";
